+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]## [0.5.4.6b1]
+## [Unreleased]## [0.5.4.6b2]
+
+## [0.5.4.6b1]- 2023-02-06
 ## Added
 - ScenarioDbManager support for custom_naming_convention of FK (and other) constraints
 ### Changed